Add ICU App-Local support on Mono
