+++ src/libraries/System.Private.CoreLib/src/System/Globalization/GlobalizationMode.Windows.cs
