+++ src/libraries/System.Private.CoreLib/src/System/Globalization/GlobalizationMode.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Diagnostics.CodeAnalysis;
-using System.Reflection;
 using System.Runtime.InteropServices;
 
 namespace System.Globalization
@@ -16,42 +15,10 @@ namespace System.Globalization
         private static bool TryGetAppLocalIcuSwitchValue([NotNullWhen(true)] out string? value) =>
             TryGetStringValue("System.Globalization.AppLocalIcu", "DOTNET_SYSTEM_GLOBALIZATION_APPLOCALICU", out value);
 
-        // GetSwitchValue calls CLRConfig first to detect if the switch is defined in the config file.
-        // if the switch is defined we just use the value of this switch. otherwise, we'll try to get the switch
-        // value from the environment variable if it is defined.
-        private static bool GetSwitchValue(string switchName, string envVariable)
-        {
-            if (!AppContext.TryGetSwitch(switchName, out bool ret))
-            {
-                string? switchValue = Environment.GetEnvironmentVariable(envVariable);
-                if (switchValue != null)
-                {
-                    ret = bool.IsTrueStringIgnoreCase(switchValue) || switchValue.Equals("1");
-                }
-            }
-
-            return ret;
-        }
-
-        private static bool TryGetStringValue(string switchName, string envVariable, [NotNullWhen(true)] out string? value)
-        {
-            value = AppContext.GetData(switchName) as string;
-            if (string.IsNullOrEmpty(value))
-            {
-                value = Environment.GetEnvironmentVariable(envVariable);
-                if (string.IsNullOrEmpty(value))
-                {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
         private static void LoadAppLocalIcu(string icuSuffixAndVersion, bool suffixWithSeparator = false)
         {
+            ReadOnlySpan<char> version;
             ReadOnlySpan<char> icuSuffix = default;
-            ReadOnlySpan<char> version = default;
 
             // Custom built ICU can have a suffix on the name, i.e: libicuucmyapp.so.67.1
             // So users would set the runtime switch as: myapp:67.1
