+++ src/mono/netcore/System.Private.CoreLib/src/System/Globalization/GlobalizationMode.Mono.cs
@@ -2,14 +2,13 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Diagnostics.CodeAnalysis;
+
 namespace System.Globalization
 {
     internal static partial class GlobalizationMode
     {
-        private static bool GetInvariantSwitchValue() =>
-            GetSwitchValue("DOTNET_SYSTEM_GLOBALIZATION_INVARIANT");
-
-        private static bool GetSwitchValue(string envVariable)
+        private static bool GetSwitchValue(string switchName, string envVariable)
         {
             bool ret = false;
             string? switchValue = Environment.GetEnvironmentVariable(envVariable);
@@ -18,6 +17,12 @@ namespace System.Globalization
                 ret = bool.IsTrueStringIgnoreCase(switchValue) || switchValue.Equals("1");
             }
             return ret;
+        }
+
+        private static bool TryGetStringValue(string switchName, string envVariable, [NotNullWhen(true)] out string? value)
+        {
+            value = Environment.GetEnvironmentVariable(envVariable);
+            return !string.IsNullOrEmpty(value);
         }
     }
 }
